Fix ambiguous Int constructor call

+++ KeyHunt-Cuda/SECP256K1.cpp
@@ -68,7 +68,7 @@ Int Secp256K1::MulShift384(const Int& k, const std::array<uint64_t, 4>& g)
         accum[6] = static_cast<uint64_t>(accum[6]);
         accum[7] = static_cast<uint64_t>(accum[7]);
 
-        Int result(0);
+        Int result(static_cast<uint64_t>(0));
         result.bits64[0] = static_cast<uint64_t>(accum[6]);
         result.bits64[1] = static_cast<uint64_t>(accum[7]);
         return result;
